Fix: Here we have implemeted many features, this is a major change. Starting from Sidebar.tsx component, the same components were imported but the content of the files has changed a lot. In the sidebar, we are using useffect hook to send a request to ./pages/api/GetChatRows.ts, we are sending a request to find all the chats that belong to a particular user, hence we are using prisma.chat.findMany, the condition is userId, so all the chats that associated with this userId will be retrieved.
Another major change happens in the NewChat.tsx component, NewChat.tsx component has an import statement that summons the ./functions/createNewChat function, this function sends an api post request to ./pages/api/CreateNewChat.ts,  CreateNewChat.ts is responsible for creating a chat, everytime we click on NewChat component, it triggers --1--> function --2--> api/ --3--> response is returned to the component. In this example, NewChat.tsx component calls a function called createNewChat, and the function calls an api called CreateNewChat.

Another major change happens in the ChatRow component. ChatRow.tsx component has an import statement that calls getMessages from ./functions/getMessages.ts, the general Idea is that we want to retrieve the messages to store some part of the messages in the chatrow boxes so when the user looks at the sidebar he will see many chatrows, but these chatrows don't hold any information, if we can retrieve the messages that was send in each of those chats, and use them inside the sidebar chatrows, so that the user can see what previous messages was sent in each chatrow before clicking on it to start interacting with chatbot, this helps the user know what was the coverstation was like from the last session it was used.